+++ package/src/components/elements/UserVideoControl.stories.tsx
@@ -215,28 +215,13 @@ export const Connected: Story<{
   variant: ButtonVariant;
   size: ButtonSize;
   state: ButtonState;
-  noAutoInitDevices: boolean;
-}> = ({ variant, size, state, noAutoInitDevices }) => (
-  <UserVideoControl
-    variant={variant}
-    size={size}
-    state={state}
-    noAutoInitDevices={noAutoInitDevices}
-  />
+}> = ({ variant, size, state }) => (
+  <UserVideoControl variant={variant} size={size} state={state} />
 );
 
 Connected.args = {
   variant: "outline",
   size: "md",
-  noAutoInitDevices: false,
-};
-
-Connected.argTypes = {
-  ...Connected.argTypes,
-  noAutoInitDevices: {
-    control: { type: "boolean" },
-    defaultValue: false,
-  },
 };
 
 Connected.decorators = [